frame-generator/requirements.txt
Flask[async]~=3.0.3
httpx~=0.27.0
aiosqlite~=0.20.0
pillow
